+++ src/main/org/firebirdsql/event/FBEventManager.java
@@ -1,52 +1,42 @@
 /*
- * Firebird Open Source JavaEE Connector - JDBC Driver
- * 
- * Copyright (C) All Rights Reserved.
- * 
- * This file was created by members of the firebird development team.
- * All individual contributions remain the Copyright (C) of those
- * individuals.  Contributors to this file are either listed here or
- * can be obtained from a source control history command.
- * 
+ * Public Firebird Java API.
+ *
  * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions
- * are met:
- *  
- *   - Redistributions of source code must retain the above copyright 
- *     notice, this list of conditions and the following disclaimer.
- *   - Redistributions in binary form must reproduce the above 
- *     copyright notice, this list of conditions and the following 
- *     disclaimer in the documentation and/or other materials provided 
- *     with the distribution.
- *   - Neither the name of the firebird development team nor the names
- *     of its contributors may be used to endorse or promote products 
- *     derived from this software without specific prior written 
- *     permission.
- * 
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS 
- * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT 
- * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS 
- * FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE 
- * COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, 
- * INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, 
- * BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS 
- * OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED 
- * AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
- * OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT 
- * OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF 
- * SUCH DAMAGE.
+ * modification, are permitted provided that the following conditions are met:
+ *    1. Redistributions of source code must retain the above copyright notice,
+ *       this list of conditions and the following disclaimer.
+ *    2. Redistributions in binary form must reproduce the above copyright
+ *       notice, this list of conditions and the following disclaimer in the
+ *       documentation and/or other materials provided with the distribution.
+ *    3. The name of the author may not be used to endorse or promote products
+ *       derived from this software without specific prior written permission.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE AUTHOR ``AS IS'' AND ANY EXPRESS OR IMPLIED
+ * WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF
+ * MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO
+ * EVENT SHALL THE AUTHOR BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+ * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO,
+ * PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS;
+ * OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY,
+ * WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR
+ * OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF
+ * ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 package org.firebirdsql.event;
 
 import org.firebirdsql.gds.EventHandle;
+import org.firebirdsql.gds.JaybirdErrorCodes;
 import org.firebirdsql.gds.impl.GDSFactory;
 import org.firebirdsql.gds.impl.GDSType;
 import org.firebirdsql.gds.ng.*;
+import org.firebirdsql.gds.ng.listeners.DefaultDatabaseListener;
 import org.firebirdsql.jdbc.FBSQLException;
+import org.firebirdsql.jdbc.FirebirdConnection;
 import org.firebirdsql.logging.Logger;
 import org.firebirdsql.logging.LoggerFactory;
 import org.firebirdsql.util.SQLExceptionChainBuilder;
 
+import java.sql.Connection;
 import java.sql.SQLException;
 import java.util.*;
 import java.util.concurrent.BlockingQueue;
@@ -57,6 +47,8 @@ import java.util.concurrent.TimeUnit;
  * An {@link org.firebirdsql.event.EventManager} implementation to listen for database events.
  *
  * @author <a href="mailto:gab_reid@users.sourceforge.net">Gabriel Reid</a>
+ * @author <a href="mailto:mrotteveel@users.sourceforge.net">Mark Rotteveel</a>
+ * @author <a href="mailto:vasiliy.yashkov@red-soft.ru">Vasiliy Yashkov</a>
  */
 public class FBEventManager implements EventManager {
 
@@ -64,8 +56,9 @@ public class FBEventManager implements EventManager {
 
     private final GDSType gdsType;
     private FbDatabase fbDatabase;
-    private final IConnectionProperties connectionProperties = new FbConnectionProperties();
-    private boolean connected = false;
+    private final IConnectionProperties connectionProperties;
+    private final EventManagerBehaviour eventManagerBehaviour;
+    private volatile boolean connected = false;
     private final Map<String, Set<EventListener>> listenerMap = Collections.synchronizedMap(new HashMap<String, Set<EventListener>>());
     private final Map<String, GdsEventHandler> handlerMap = Collections.synchronizedMap(new HashMap<String, GdsEventHandler>());
     private final BlockingQueue<DatabaseEvent> eventQueue = new LinkedBlockingQueue<>();
@@ -80,16 +73,68 @@ public class FBEventManager implements EventManager {
 
     public FBEventManager(GDSType gdsType) {
         this.gdsType = gdsType;
+        connectionProperties = new FbConnectionProperties();
+        eventManagerBehaviour = new DefaultEventManagerBehaviour();
     }
 
+    /**
+     * Constructs an event manager using an existing connection.
+     *
+     * @param connection Connection that unwraps to {@link FirebirdConnection}.
+     */
+    private FBEventManager(Connection connection) throws SQLException {
+        this.fbDatabase = connection.unwrap(FirebirdConnection.class).getFbDatabase();
+        gdsType = null;
+        connectionProperties = fbDatabase.getConnectionProperties().asImmutable();
+        // NOTE In this implementation, we don't take into account pooled connections that might be closed while
+        //  the FbDatabase instance remains in use. This means that at the moment, it is possible that the event manager
+        //  can remain in use for longer than the Connection.
+        fbDatabase.addDatabaseListener(new DefaultDatabaseListener() {
+            @Override
+            public void detaching(FbDatabase database) {
+                try {
+                    if (!isConnected()) return;
+                    try {
+                        disconnect();
+                    } catch (SQLException e) {
+                        log.error("Exception on disconnect of event manager on connection detaching.", e);
+                    }
+                } finally {
+                    database.removeDatabaseListener(this);
+                    fbDatabase = null;
+                }
+            }
+        });
+        eventManagerBehaviour = new ManagedEventManagerBehaviour();
+    }
+
+    /**
+     * Creates an {@link EventManager} for a connection.
+     * <p>
+     * The created event manager does not allow setting the properties and will instead
+     * throw {@link UnsupportedOperationException} for the setters.
+     * </p>
+     * <p>
+     * The returned instance is not necessarily an implementation of {@link FBEventManager}.
+     * </p>
+     *
+     * @param connection
+     *         A connection that unwraps to {@link org.firebirdsql.jdbc.FirebirdConnection}
+     * @return An event manager
+     * @throws SQLException
+     *         When {@code connection} does not unwrap to {@link org.firebirdsql.jdbc.FirebirdConnection}
+     * @since 3.0.7
+     */
+    public static EventManager createFor(Connection connection) throws SQLException {
+        return new FBEventManager(connection);
+    }
+
+    @Override
     public void connect() throws SQLException {
         if (connected) {
             throw new IllegalStateException("Connect called while already connected");
         }
-        final FbDatabaseFactory databaseFactory = GDSFactory.getDatabaseFactoryForType(gdsType);
-
-        fbDatabase = databaseFactory.connect(connectionProperties);
-        fbDatabase.attach();
+        eventManagerBehaviour.connectDatabase();
         connected = true;
 
         eventDispatcher = new EventDispatcher();
@@ -98,6 +143,14 @@ public class FBEventManager implements EventManager {
         dispatchThread.start();
     }
 
+    @Override
+    public void close() throws SQLException {
+        if (connected) {
+            disconnect();
+        }
+    }
+
+    @Override
     public void disconnect() throws SQLException {
         if (!connected) {
             throw new IllegalStateException("Disconnect called while not connected");
@@ -118,7 +171,7 @@ public class FBEventManager implements EventManager {
                 handlerMap.clear();
                 listenerMap.clear();
                 try {
-                    fbDatabase.close();
+                    eventManagerBehaviour.disconnectDatabase();
                 } catch (SQLException e) {
                     chain.append(e);
                 }
@@ -126,57 +179,71 @@ public class FBEventManager implements EventManager {
             }
         } finally {
             eventDispatcher.stop();
-
+            dispatchThread.interrupt();
             // join the thread and wait until it dies
             try {
                 dispatchThread.join();
             } catch (InterruptedException ex) {
                 chain.append(new FBSQLException(ex));
+            } finally {
+                eventDispatcher = null;
+                dispatchThread = null;
             }
         }
         if (chain.hasException()) throw chain.getException();
     }
 
+    @Override
     public boolean isConnected() {
         return connected;
     }
 
+    @Override
     public void setUser(String user) {
         connectionProperties.setUser(user);
     }
 
+    @Override
     public String getUser() {
         return connectionProperties.getUser();
     }
 
+    @Override
     public void setPassword(String password) {
         connectionProperties.setPassword(password);
     }
 
+    @Override
     public String getPassword() {
         return connectionProperties.getPassword();
     }
 
+    @Override
     public void setDatabase(String database) {
         connectionProperties.setDatabaseName(database);
     }
 
+    @Override
     public String getDatabase() {
         return connectionProperties.getDatabaseName();
     }
 
+    @Override
     public String getHost() {
         return connectionProperties.getServerName();
     }
 
+    @Override
     public void setHost(String host) {
         connectionProperties.setServerName(host);
     }
 
+    @Override
     public int getPort() {
         return connectionProperties.getPortNumber();
     }
 
+    @Override
     public void setPort(int port) {
         connectionProperties.setPortNumber(port);
     }
@@ -201,32 +268,13 @@ public class FBEventManager implements EventManager {
         connectionProperties.setDbCryptConfig(dbCryptConfig);
     }
 
-    /**
-     * Get the time in milliseconds, after which the async thread will exit from the {@link Object#wait(long)} method
-     * and check whether it was stopped or not.
-     * <p>
-     * Default value is 1000 (1 second).
-     * </p>
-     *
-     * @return wait timeout in milliseconds
-     */
-    @SuppressWarnings("UnusedDeclaration")
+    @Override
     public long getWaitTimeout() {
         return waitTimeout;
     }
 
-    /**
-     * Set the time in milliseconds, after which the async thread will exit from the {@link Object#wait(long)} method
-     * and check whether it was stopped or not.
-     * <p>
-     * Default value is 1000 (1 second).
-     * </p>
-     *
-     * @param waitTimeout
-     *         wait timeout in milliseconds
-     */
-    @SuppressWarnings("UnusedDeclaration")
-    public synchronized void setWaitTimeout(long waitTimeout) {
+    @Override
+    public void setWaitTimeout(long waitTimeout) {
         this.waitTimeout = waitTimeout;
     }
 
@@ -294,9 +342,53 @@ public class FBEventManager implements EventManager {
     private void unregisterListener(String eventName) throws SQLException {
         GdsEventHandler handler = handlerMap.get(eventName);
         try {
-            handler.unregister();
+            if (handler != null) handler.unregister();
         } finally {
             handlerMap.remove(eventName);
+        }
+    }
+
+    private interface EventManagerBehaviour {
+        void connectDatabase() throws SQLException;
+        void disconnectDatabase() throws SQLException;
+    }
+
+    /**
+     * Default behaviour where the event manager owns the connection.
+     */
+    private class DefaultEventManagerBehaviour implements EventManagerBehaviour {
+
+        @Override
+        public void connectDatabase() throws SQLException {
+            FbDatabaseFactory databaseFactory = GDSFactory.getDatabaseFactoryForType(gdsType);
+            fbDatabase = databaseFactory.connect(connectionProperties);
+            fbDatabase.attach();
+        }
+
+        @Override
+        public void disconnectDatabase() throws SQLException {
+            fbDatabase.close();
+        }
+    }
+
+    /**
+     * Behaviour where the lifetime of the connection used by the event manager is managed elsewhere.
+     */
+    private class ManagedEventManagerBehaviour implements EventManagerBehaviour {
+
+        @Override
+        public void connectDatabase() throws SQLException {
+            // using existing connection
+            if (fbDatabase == null) {
+                // fbDatabase has already detached
+                throw FbExceptionBuilder.forException(JaybirdErrorCodes.jb_notConnectedToServer)
+                        .toFlatSQLException();
+            }
+        }
+
+        @Override
+        public void disconnectDatabase() {
+            // fbDatabase will be closed where it was opened
         }
     }
 
@@ -323,12 +415,15 @@ public class FBEventManager implements EventManager {
             cancelled = true;
         }
 
+        @Override
         public synchronized void eventOccurred(EventHandle eventHandle) {
             if (!cancelled) {
                 try {
                     fbDatabase.countEvents(eventHandle);
                 } catch (SQLException e) {
-                    log.warn("Exception processing event counts", e);
+                    String message = "Exception processing event counts";
+                    log.warn(message + ": " + e + "; see debug level for stacktrace");
+                    log.debug(message, e);
                 }
 
                 if (initialized && !cancelled) {
@@ -340,7 +435,9 @@ public class FBEventManager implements EventManager {
                 try {
                     register();
                 } catch (SQLException e) {
-                    log.warn("Exception registering for event", e);
+                    String message = "Exception registering for event";
+                    log.warn(message + ": " + e + "; see debug level for stacktrace");
+                    log.debug(message, e);
                 }
             }
         }
@@ -348,14 +445,14 @@ public class FBEventManager implements EventManager {
 
     class EventDispatcher implements Runnable {
 
-        private volatile boolean running = false;
+        private volatile boolean running = true;
 
         public void stop() {
             running = false;
         }
 
+        @Override
         public void run() {
-            running = true;
             DatabaseEvent event;
             while (running) {
                 try {
@@ -388,6 +485,7 @@ class OneTimeEventListener implements EventListener {
         this.lock = lock;
     }
 
+    @Override
     public void eventOccurred(DatabaseEvent event) {
         if (eventCount == -1) {
             eventCount = event.getEventCount();
@@ -413,14 +511,17 @@ class DatabaseEventImpl implements DatabaseEvent {
         this.eventCount = eventCount;
     }
 
+    @Override
     public int getEventCount() {
         return this.eventCount;
     }
 
+    @Override
     public String getEventName() {
         return this.eventName;
     }
 
+    @Override
     public String toString() {
         return "DatabaseEvent['" + eventName + " * " + eventCount + "]";
     }